libcrypt: fix compiler warnings about variable types

+++ lib/libcrypt/crypt.c
@@ -32,6 +32,7 @@ __FBSDID("$FreeBSD$");
 
 #include <sys/types.h>
 
+#include <libutil.h>
 #include <string.h>
 #include <unistd.h>
 
@@ -110,7 +111,7 @@ crypt_r(const char *passwd, const char *salt, struct crypt_data *data)
 	const struct crypt_format *cf;
 	int (*func)(const char *, const char *, char *);
 #ifdef HAS_DES
-	int len;
+	size_t len;
 #endif
 
 	for (cf = crypt_formats; cf->name != NULL; ++cf)
